+++ src/icon/__tests__/icon.spec.js
@@ -1,10 +1,13 @@
-import makeGetWrapper from '../../utils/unit-testing/get-enzyme-wrapper';
+import { makeGetEnzymeWrapper } from '../../utils/unit-testing/get-enzyme-wrapper';
 
 import { StyledIcon } from '../icon.styled';
 import Icon from '../index';
 
 describe('Icon', () => {
-    const getWrapper = makeGetWrapper(Icon, { size: 'medium', icon: 'mdIcon' });
+    const getWrapper = makeGetEnzymeWrapper(Icon, {
+        size: 'medium',
+        icon: 'mdIcon',
+    });
 
     describe('rendering', () => {
         it('should render a StyledIcon', () => {