+++ src/utils/unit-testing/get-enzyme-wrapper/get-enzyme-wrapper.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-function makeGetEnzymeWrapper(Component, defaultProps = {}) {
+export function makeGetEnzymeWrapper(Component, defaultProps = {}) {
     return function getEnzymeWrapper(props = {}, depth = mount) {
         const mergedProps = { ...defaultProps, ...props };
         return depth(<Component {...mergedProps} />);
     };
 }
-
-export default makeGetEnzymeWrapper;